at last got the facebook and google plus sharing to work!
