+++ css/property-reviews/social.css
@@ -25,25 +25,29 @@
 		text-decoration: none;
 	color: #FFF;
 }
-
-
-.fb-share {
+.gplus-share {
 	text-align: center;
 	vertical-align: middle;
-	height: 23px;
+	height: 35px;
 	width: 35px;
 	font-family: Tahoma, Geneva, sans-serif;
 	color: #FFF;
 	font-size: 20px;
-	background-color: #4267B2;
 	position: absolute;
-	left: 150px;
+	left: 140px;
 	top: 0px;
-	border: thin solid #4267B2;
 	border-radius:4px;
-	padding-top: 3px;
 }
-
-.fb-share:hover {
-	background-color: #365899;
+.in-share {
+	text-align: center;
+	vertical-align: middle;
+	height: 35px;
+	width: 35px;
+	font-family: Tahoma, Geneva, sans-serif;
+	color: #FFF;
+	font-size: 20px;
+	position: absolute;
+	left: 180px;
+	top: 0px;
+	border-radius:4px;
 }
